requirements.txt
pymupdf
PyPDF2
pillow
rembg[cpu,cli]
onnxruntime
pytesseract
